Generate PCB assembly files
Generate assembly files for JLCPCB - BOM and CPL. Using existing LCSC part numbers (which may not be in the JLC part library), and positions/rotations are completely unverified, and of course many of the components are simply not available in any form in the JLC part library (e.g. TMC6300). Fixes #23.

Also changed LCD backlight resistor from 22 ohm to 20 ohm and added some additional part numbers for reference.